refactor: move subs menu to isolated component
